requirements.txt
opencv-python
numpy
Pillow
deepface
torch
tensorflow
onx
onnx2keras  